Add Offline mode messages
