+++ android/source/VideoLocker/res/layout/activity_find_courses.xml
@@ -31,16 +31,7 @@
 
         <org.edx.mobile.view.custom.ETextView
             android:id="@+id/offline_mode_message"
-            style="@style/bold_text"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/find_courses_offline_text"
-            android:padding="10dp"
-            android:layout_centerInParent="true"
-            android:textColor="@color/empty_list_text"
-            android:textSize="20sp"
-            android:visibility="gone"
-            android:gravity="center" />
+            style="@style/offline_mode_message"/>
 
     </RelativeLayout>
