+++ android/source/VideoLocker/res/values/text_styles.xml
@@ -137,4 +137,16 @@
         <item name="android:layout_height">wrap_content</item>
         <item name="android:gravity">center</item>
      </style>
+
+    <style name="offline_mode_message" parent="@style/bold_text">
+        <item name="android:textSize">20sp</item>
+        <item name="android:layout_width">match_parent</item>
+        <item name="android:layout_height">wrap_content</item>
+        <item name="android:gravity">center</item>
+        <item name="android:text">@string/find_courses_offline_text</item>
+        <item name="android:padding">10dp</item>
+        <item name="android:layout_centerInParent">true</item>
+        <item name="android:textColor">@color/empty_list_text</item>
+        <item name="android:visibility">gone</item>
+    </style>
 </resources>
