+++ android/source/VideoLocker/src/org/edx/mobile/base/FindCoursesBaseActivity.java
@@ -1,0 +1,119 @@
+package org.edx.mobile.base;
+
+import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
+import android.view.View;
+import android.widget.ProgressBar;
+
+import org.edx.mobile.R;
+import org.edx.mobile.util.AppConstants;
+import org.edx.mobile.util.NetworkUtil;
+import org.edx.mobile.view.custom.ETextView;
+
+public class FindCoursesBaseActivity extends BaseFragmentActivity {
+
+    private View offlineBar;
+
+    @Override
+    protected void onCreate(Bundle arg0) {
+        super.onCreate(arg0);
+
+        offlineBar = findViewById(R.id.offline_bar);
+        if (!(NetworkUtil.isConnected(this))) {
+            AppConstants.offline_flag = true;
+            invalidateOptionsMenu();
+            showOfflineMessage();
+            if(offlineBar!=null){
+                offlineBar.setVisibility(View.VISIBLE);
+            }
+        }
+    }
+
+    @Override
+    protected void onOnline() {
+        offlineBar.setVisibility(View.GONE);
+        hideOfflineMessage();
+        invalidateOptionsMenu();
+    }
+
+    @Override
+    protected void onOffline() {
+        offlineBar.setVisibility(View.VISIBLE);
+        showOfflineMessage();
+        hideLoadingProgress();
+        invalidateOptionsMenu();
+    }
+
+    @Override
+    protected void onStart() {
+        super.onStart();
+        setTitle(getString(R.string.find_courses_title));
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        super.onCreateOptionsMenu(menu);
+
+        //TODO - Remove checkbox code after the PR for removing checkbox is merged with master
+        MenuItem checkBox_menuItem = menu.findItem(R.id.delete_checkbox);
+        checkBox_menuItem.setVisible(false);
+
+        //Hide the download progress from Action bar
+        MenuItem menuItem = menu.findItem(R.id.progress_download);
+        menuItem.setVisible(false);
+
+        return true;
+    }
+
+    @Override
+    public boolean onPrepareOptionsMenu(Menu menu) {
+        //Hide the download progress from Action bar.
+        //This has to be called in onCreateOptions as well
+        MenuItem menuItem = menu.findItem(R.id.progress_download);
+        menuItem.setVisible(false);
+        return true;
+    }
+
+    /**
+     * This function shows the offline mode message
+     */
+    private void showOfflineMessage(){
+        ETextView offlineModeTv = (ETextView) findViewById(R.id.offline_mode_message);
+        if(offlineModeTv!=null){
+            offlineModeTv.setVisibility(View.VISIBLE);
+        }
+    }
+
+    /**
+     * This function hides the offline mode message
+     */
+    private void hideOfflineMessage(){
+        ETextView offlineModeTv = (ETextView) findViewById(R.id.offline_mode_message);
+        if(offlineModeTv!=null){
+            offlineModeTv.setVisibility(View.GONE);
+        }
+    }
+
+    /**
+     * This function shows the loading progress wheel
+     * Show progress wheel while loading the web page
+     */
+    private void showLoadingProgress(){
+        ProgressBar progressWheel = (ProgressBar) findViewById(R.id.progress_spinner);
+        if(progressWheel!=null){
+            progressWheel.setVisibility(View.VISIBLE);
+        }
+    }
+
+    /**
+     * This function hides the loading progress wheel
+     * Hide progress wheel after the web page completes loading
+     */
+    private void hideLoadingProgress(){
+        ProgressBar progressWheel = (ProgressBar) findViewById(R.id.progress_spinner);
+        if(progressWheel!=null){
+            progressWheel.setVisibility(View.GONE);
+        }
+    }
+}
