+++ android/source/VideoLocker/src/org/edx/mobile/view/CourseInfoActivity.java
@@ -3,17 +3,21 @@ package org.edx.mobile.view;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuItem;
+import android.view.View;
 
 import org.edx.mobile.R;
 import org.edx.mobile.base.BaseFragmentActivity;
+import org.edx.mobile.base.FindCoursesBaseActivity;
 import org.edx.mobile.module.analytics.ISegment;
+import org.edx.mobile.util.AppConstants;
+import org.edx.mobile.util.NetworkUtil;
 
-public class CourseInfoActivity extends BaseFragmentActivity {
+public class CourseInfoActivity extends FindCoursesBaseActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_find_course_info);
+        super.onCreate(savedInstanceState);
 
         try{
             segIO.screenViewsTracking(ISegment.Values.COURSE_INFO_SCREEN);
@@ -22,18 +26,4 @@ public class CourseInfoActivity extends BaseFragmentActivity {
         }
     }
 
-    @Override
-    protected void onStart() {
-        super.onStart();
-        setTitle(getString(R.string.find_courses_title));
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        super.onCreateOptionsMenu(menu);
-
-        MenuItem checkBox_menuItem = menu.findItem(R.id.delete_checkbox);
-        checkBox_menuItem.setVisible(false);
-        return true;
-    }
 }
